do a bunch of renaming
