+++ src/sql/platform/restore/common/restoreServiceImpl.ts
@@ -19,8 +19,8 @@ import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';
 import { ProviderConnectionInfo } from 'sql/platform/connection/common/providerConnectionInfo';
 import * as Utils from 'sql/platform/connection/common/utils';
 import { IObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
-import { ITaskService } from 'sql/platform/taskHistory/common/taskService';
-import { TaskStatus, TaskNode } from 'sql/workbench/parts/taskHistory/common/taskNode';
+import { ITaskService } from 'sql/platform/tasks/common/tasksService';
+import { TaskStatus, TaskNode } from 'sql/platform/tasks/common/tasksNode';
 import * as TelemetryKeys from 'sql/platform/telemetry/telemetryKeys';
 import * as TelemetryUtils from 'sql/platform/telemetry/telemetryUtilities';
 import { IConnectionManagementService } from 'sql/platform/connection/common/connectionManagement';
