+++ src/sql/workbench/parts/tasks/browser/media/tasksPanel.css
