+++ src/sql/workbench/parts/tasks/browser/tasksRenderer.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ITree, IRenderer } from 'vs/base/parts/tree/browser/tree';
-import { TaskNode, TaskStatus } from 'sql/workbench/parts/taskHistory/common/taskNode';
+import { TaskNode, TaskStatus } from 'sql/platform/tasks/common/tasksNode';
 import * as dom from 'vs/base/browser/dom';
 import { localize } from 'vs/nls';
 import * as Utils from 'sql/platform/connection/common/utils';
