+++ src/sql/workbench/parts/tasks/common/tasksAction.ts
@@ -5,8 +5,8 @@
 
 import { localize } from 'vs/nls';
 import { Action } from 'vs/base/common/actions';
-import { ITaskService } from 'sql/platform/taskHistory/common/taskService';
-import { TaskNode } from 'sql/workbench/parts/taskHistory/common/taskNode';
+import { ITaskService } from 'sql/platform/tasks/common/tasksService';
+import { TaskNode } from 'sql/platform/tasks/common/tasksNode';
 import { IQueryEditorService } from 'sql/workbench/services/queryEditor/common/queryEditorService';
 import Severity from 'vs/base/common/severity';
 import { IErrorMessageService } from 'sql/platform/errorMessage/common/errorMessageService';
