+++ src/vs/workbench/workbench.main.ts
@@ -183,7 +183,7 @@ import { ICredentialsService, CredentialsService } from 'sql/platform/credential
 import { ISerializationService, SerializationService } from 'sql/platform/serialization/common/serializationService';
 import { IMetadataService, MetadataService } from 'sql/platform/metadata/common/metadataService';
 import { IObjectExplorerService, ObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
-import { ITaskService, TaskService } from 'sql/platform/taskHistory/common/taskService';
+import { ITaskService, TaskService } from 'sql/platform/tasks/common/tasksService';
 import { IQueryModelService } from 'sql/platform/query/common/queryModel';
 import { QueryModelService } from 'sql/platform/query/common/queryModelService';
 import { IQueryEditorService } from 'sql/workbench/services/queryEditor/common/queryEditorService';
